Prevent inconsistency due to 'DOWN' received after nodeup
In the case of a connection loss and a subsequent reconnect,
a {'DOWN', _,_,_ noconnection} message, due to a monitor
between peers over the connection, may arrive after a {nodeup, _}
message. This can cause an inconsistency where a node considers
the other node down when it actually is up and connected.

By clearing information about the other node when {nodeup, _}
arrives we ensure that we save valid information when the other
node responds.

We only clear information about the other node on reception of
a {'DOWN', Mref,_,_ noconnection} when Mref corresponds to the
Mref in the saved information. If Mref corresponds to saved
information we know we have not received a {nodeup, _} since
we saved the information.